Integration of twitter mock stream with microservice

+++ app-config-data/src/main/java/com/jdutton/poc/config/RetryConfigData.java
@@ -12,7 +12,7 @@ public class RetryConfigData {
 	
 	private Long initialIntervalMs;
 	private Long maxIntervalMs;
-	private Integer multiplier;
+	private Double multiplier;
 	private Integer maxAttempts;
 	private Long sleepTimeMs;
 }
